Move save and load to App class
